+++ src/utils/debounce.ts
@@ -1,0 +1,13 @@
+const debounce = (func: () => void, delay: number) => {
+  let timer: NodeJS.Timeout;
+
+  return () => {
+    clearTimeout(timer);
+
+    timer = setTimeout(() => {
+      func();
+    }, delay);
+  };
+};
+
+export { debounce };
